# Add bulk api for checking networks if they contain an ip address