Improves performance and reduces allocations in the scenario for checking one address against many `IPNetwork` instances. Effectively saves ~36 bytes for each network that is checked, how many nanoseconds saved per invocation is harder to pin down. Savings come purely from just caching the numerical value of the IPAddress (benchmarks are ipv4 values only, ipv6 may have larger gains).

|       Method |       Mean |    Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |-----------:|---------:|----------:|------:|-------:|------:|------:|----------:|
|   Contains x 1 |   458.5 ns |  7.92 ns |   7.02 ns |  1.00 | 0.0553 |     - |     - |     292 B |
|              |            |          |           |       |        |       |       |           |
|  Original Contains x 2 |   920.8 ns | 18.15 ns |  17.83 ns |  1.00 | 0.1106 |     - |     - |     585 B |
| Optimized Contains x 2 |   830.9 ns | 15.73 ns |  14.72 ns |  1.00 | 0.1040 |     - |     - |     549 B |
|              |            |          |           |       |        |       |       |           |
|  Original Contains x 10 | 4,377.9 ns | 31.78 ns |  26.54 ns |  1.00 | 0.5264 |     - |     - |    2764 B |
| Optimized Contains x 10 | 3,741.4 ns | 69.03 ns | 117.23 ns |  1.00 | 0.4616 |     - |     - |    2440 B |