@darrenkopp 2021-02-17
I went with `Contains` as the method name, but perhaps `ContainsAny` would be more appropriate?