@lduchosal 2021-03-03
Thanks @darrenkopp ! 

A few remarks : 

- [ ]  Array implements IEnumerable. The whole method is not needed and is duplicated code.
        Can we just remove method :  Contains(IPNetwork[] networks, IPAddress address) ?
     
- [ ]  Looks like it's business belongs to IPNetworkCollection rather than IPNetwork
       Can we move this Contains method  to IPNetworkCollection  ?

- [ ] Unit test : can you add some unit test ? i'll do if time permits.