@darrenkopp 2021-03-03
@lduchosal
> Array implements IEnumerable. The whole method is not needed and is duplicated code.

While true, you'll see allocations + performance decrease by stepping down to `IEnumerable<T>`, but yes it would work. The compiler does some specific optimizations for `T[]` instances but I believe it does that for `IList<T>` as well, which would then work optimially for both arrays and `List<T>` instances. Will investivate, but I think that'd be the only two overloads for optimal behavior / duplication tradeoff.

I should have made the method more general, what can be seen here is purely all of the call sites that my code needed 👼 

> Looks like it's business belongs to IPNetworkCollection rather than IPNetwork. Can we move this Contains method to IPNetworkCollection ?

Yeah I can move it.

> Unit test : can you add some unit test ? i'll do if time permits.

Yeah I can add some.